+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+1. Adds `Entity::Fund` `active?` and `default?` interrogatives
+
 ## [4.0.1] - 2024-09-13
 
 This version adds support for Ruby 3.3.5. No real changes were required. We just needed to add some gems to the gemspec